requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
bcrypt==3.2.0
    # via passlib
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via uvicorn
cryptography==36.0.1
    # via python-jose
dnspython==2.2.1
    # via email-validator
ecdsa==0.17.0
    # via python-jose
email-validator==1.1.3
    # via fastapi
fastapi[all]==0.75.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
itsdangerous==2.1.0
    # via fastapi
jinja2==3.0.3
    # via fastapi
markupsafe==2.1.0
    # via jinja2
orjson==3.6.7
    # via fastapi
passlib[bcrypt]==1.7.4
    # via -r requirements.in
psycopg2-binary[all]==2.9.3
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via fastapi
python-decouple==3.6
    # via -r requirements.in
python-dotenv==0.19.2
    # via uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via fastapi
pyyaml==5.4.1
    # via
    #   fastapi
    #   uvicorn
requests==2.27.1
    # via fastapi
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   bcrypt
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy[all]==1.4.32
    # via -r requirements.in
starlette==0.17.1
    # via fastapi
typing-extensions==4.1.1
    # via pydantic
ujson==4.3.0
    # via fastapi
urllib3==1.26.8
    # via requests
uvicorn[standard]==0.15.0
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.2
    # via uvicorn
